Add new wlctl_led driver, refs #9413
Feature primarily aimed at DG150v2 that have the "Wireless" led connected to
the wifi board. The driver works by spawning the Broadcom wlctl utility.
For this reason it bocomes quite slow to react (up to 1s) and thus there
is no point in using FLASH_FAST for leds using this driver.