Workaround for ghost gpio press directly after start
On platforms with gpio buttons we have seen that sometimes buttons
get pressed directly after start and thus disabling the wifi. This
change implements a dummy read to flush out the bad states. Also
implemented is a check if a button is pressed directly 4 seconds
after start. If that is the case the press is logged to the syslog
and then ignored.